Improve clarity of your doc
Please look up how to use which ("If the sentence doesn’t need the clause that the word in question is connecting, use which. If it does, use that."), and strip down your language to the basics. You're using a LOT of passive voice, the phrases "such that" and "as well as" really don't belong in technical documentation, and the word should does not belong in advice unless it is a recommendation, and even then, you'd want to put a voice behind the recommendation. You're writing this like an authoritative text, so stick with that voice, or get more personal and say that you recommend it personally. I feel like you were trolling me by including the word "synergizes". There are some cases where my rewrite might deviate from your intended meaning so take that as an indication that your intended meaning was unclear. Be consistent about your capitalizations of Docker the product and the services as well.